Semantic change to make guide more clear.
I was following this guide and saw that the sentence "Put your device in DFU Mode before running this command" was put underneath the command (this is for Step 6 on Linux and Step 3 on macOS). 

I almost missed this step as it was not highlighted and because it came **after** the step itself.

I just made a small change to bold the part "before running this command" to draw more attention to it.

Thanks.