Add unofficial version of Taurine that actually works for M1 iPad Pro's
To simplify this, basically due to various factors, the implementation of kfd in the official latest release of Taurine is, at best, extremely unreliable with M1 iPad Pro's, and generally tends to outright not work for people.

Due to my interesting purchasing decisions, I happened to acquire an M1 iPad Pro on 14.6, and ended up modifying the kfd exploit and building a modified Taurine ipa that works fairly reliably for M1 iPad Pro's.

Am I trying to get this merged into mainline? Yes (https://github.com/Odyssey-Team/Taurine/pull/49) but I don't have a lot of hope in it being merged, due to how the changes needed are impractical to implement for non-M1 devices.

Oh, and before anybody asks, yes, somebody else besides me tested this, it works fine.