Recommend Dopamine 2.1.7 for arm64e 15.x
This is a bit of an elaborate thing, so I'm going to leave an extraordinarily long explanation here detailing this commit.

Basically, as a side effect of the removal of dyld_interpose's with Dopamine 2.2 (which are a net benefit in most cases), forkfix now has to inject systemwide.

Unfortunately, arm64e iOS 15 has a stock bug which will lead to random panics commonly referred to as "Spinlock Panics", which are basically unfixable (we'd need a proper KTRR Bypass), but is linked to the number of C function hooks tweaks or the jailbreak use (stock iOS doesn't use C function hooks, so that's why stock iOS never experiences this).

The implication is that having forkfix injected systemwide results in an increase in spinlock panics - sometimes by a meaningful amount.

Based on what I've observed with other people (through the r/jailbreak subreddit and Discord), *at best*, Dopamine 2.2+ barely makes the rate at which they occur worse. But, *at worst*, 2.1.7 can panic substantially more often for some users than Dopamine 2.2+ (with identical setups).

Based on this, the most sensible path for now is to recommend 2.1.7 for arm64e 15.x, as the changes that Dopamine 2.2+ make aren't worth the potentially substantial increase in random panics for arm64e 15.x

This is not a final decision - this has a decent enough chance to get reverted based on future Dopamine updates or based on more user experiences with arm64e 15.x - however the information right now suggests that this is a net positive decision.