Recommend other options on TrollMisaka page
TrollMisaka technically gets priority because it's the only option for A8(X) iPads on 15.7.2-15.8.3 that does not have a jailbreak requirement *or* a *hard* computer requirement, but it's also not a good method overall.

Therefore, let's recommend either TrollRestore or Dopamine -> TrollHelper in-line, depending on what the person is trying or willing to do, and limit the number of people doing TrollMisaka who are able/willing to do better methods that have more requirements/prerequisites.