Update installing-socket.md
From what I have just personally experienced and have found through my research and (still ongoing support ticket on the LegacyJailbreak Discord, if you have input for me or help, please LMK as if there is a workaround for this, this change isnt necessary, and this PR can be closed,) this note is incredibly necessary for software preservation, as IPA extraction NEEDS these tools, and if people are not aware of this, they may need to reset their device to undo the JB as Socket does not let you restore rootFS, and a restore could mean losing the app, which is the whole reason why they're jailbreaking. This should be an issue for Socket to fix, however, it's evident for now it has to just be a warning.