[MAJOR] Switch from Sideloadly to PlumeImpactor (mostly)
This commit switches the primary sideloading utility for most guides from Sideloadly to PlumeImpactor.

While there are a good number of reasons to switch to PlumeImpactor, the main two reasons behind switching to PlumeImpactor are:
- Native Linux Support
- Being OSS (Sideloadly has no public source)

To any average user reading this commit message: Sideloadly still works perfectly fine (this isn't like 4 years ago where AltStore/AltDeploy/whatever else was persistently annoying), however when looking at it from an overall guide perspective, it makes more sense both now and in the future to switch to PlumeImpactor as the default sideloading utility.

With that being said, the following guides have not been switched to PlumeImpactor at this time:
- unc0ver (base guide only, sideloading u0 directly with PlumeImpactor triggers u0's DRM that didn't/doesn't even accomplish anything productive)
- Any jailbreaks that support iOS 8 (PlumeImpactor only supports iOS 9 and later as of now, iOS 8 in general is finicky for stuff to support)
- tvOS Jailbreaks/Guides (PlumeImpactor does not support tvOS at this time)

The existing Sideloadly guides will be archived and will remain for a little bit in case anything explodes (as PlumeImpactor is still relatively new), but for all intents and purposes the archived Sideloadly guides are not being actively maintained and will likely be removed in a few weeks.